requirements.txt
pygame
opencv-python
anthropic
python-dotenv